Undo numpy image conversion to float32 unecessary

+++ datasets_processor/ImageFastDataset.py
@@ -233,7 +233,7 @@ class ImageFastDataset(Dataset):
 
     def get_image_from_idx(self, idx):
         image = self._load_raw_image(self._raw_idx[idx])
-        image = image.astype(np.float32)
+        # image = image.astype(np.float32)
         return image
 
     def _get_raw_labels(self):
